move discover norm rule to post-normalizer dir
